Hang a bit more flesh on the bones of the ScriptEngineFactory demo.
Create the src/main/resources/META-INF/services folder with a one-line config file.
Implement some of the methods of the ScriptEngine{,Factory}. With a toy main program,
this actually finds the engine and gets into its eval() method, which is just a stub for now,
and the main program gets back a (hard-coded for now) return value from the script engine.